pyproject.toml
[project]
name = "cloud-stroage-and-sharing-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Miku-79",email = "deepaknoob957@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
